+++ qcop/adapters/terachem.py
@@ -1,6 +1,7 @@
+from pathlib import Path
 from typing import Callable, Optional, Tuple
 
-from qcio import CalcType, ProgramInput, SinglePointResults
+from qcio import CalcType, ProgramInput, SinglePointOutput, SinglePointResults
 from qcparse import parse_results
 from qcparse.parsers.terachem import parse_version_string
 
@@ -16,6 +17,7 @@ class TeraChemAdapter(ProgramAdapter):
     supported_calctypes = [CalcType.energy, CalcType.gradient, CalcType.hessian]
     program = "terachem"
     padding = 20  # padding between keyword and value in tc.in
+    xyz_filename = "geom.xyz"
 
     def program_version(self, stdout: Optional[str] = None) -> str:
         """Get the program version.
@@ -42,8 +44,7 @@ class TeraChemAdapter(ProgramAdapter):
             Filename of input file (tc.in).
         """
         # Write molecule to disk
-        xyz_filename = "geom.xyz"
-        inp_obj.molecule.save(xyz_filename)
+        inp_obj.molecule.save(self.xyz_filename)
 
         # Write input file
         inp_filename = "tc.in"
@@ -56,7 +57,7 @@ class TeraChemAdapter(ProgramAdapter):
                 calctype = inp_obj.calctype
             f.write(f"{'run':<{self.padding}} {calctype}\n")
             # Molecule
-            f.write(f"{'coordinates':<{self.padding}} {xyz_filename}\n")
+            f.write(f"{'coordinates':<{self.padding}} {self.xyz_filename}\n")
             f.write(f"{'charge':<{self.padding}} {inp_obj.molecule.charge}\n")
             f.write(f"{'spinmult':<{self.padding}} {inp_obj.molecule.multiplicity}\n")
             # Model
@@ -85,11 +86,12 @@ class TeraChemAdapter(ProgramAdapter):
 
     # TODO: Need command line options for TeraChem e.g., -g 1 for GPUs MAYBE?
     # Try using it for a while without and see what roadblocks we run into
-    def _compute(
+    def compute_results(
         self,
         inp_obj,
         update_func: Optional[Callable] = None,
         update_interval: Optional[float] = None,
+        **kwargs,
     ) -> Tuple[SinglePointResults, str]:
         """Execute TeraChem on the given input.
 
@@ -106,3 +108,66 @@ class TeraChemAdapter(ProgramAdapter):
         stdout = execute_subprocess(self.program, [tc_in], update_func, update_interval)
         parsed_output = parse_results(stdout, self.program, "stdout")
         return parsed_output, stdout
+
+    def collect_wfn(self, output: SinglePointOutput) -> None:
+        """Append wavefunction data to the output.
+
+        Args:
+            output: SinglePointOutput object on which to append wavefunction data.
+
+        Returns:
+            None. Modifies the output object in place.
+        """
+
+        # Naming conventions from TeraChem uses xyz filename as scratch dir postfix
+        scr_postfix = self.xyz_filename.split(".")[0]
+
+        # Wavefunction filenames
+        wfn_filenames = ("c0", "ca0", "cb0")
+        wfn_paths = [Path(f"scr.{scr_postfix}/{fn}") for fn in wfn_filenames]
+        if not any(wfn_path.exists() for wfn_path in wfn_paths):
+            raise AdapterInputError(
+                program=self.program,
+                message=f"No wavefunction files found in {Path.cwd()}",
+            )
+        for wfn_path in wfn_paths:
+            if wfn_path.exists():
+                output.files[str(wfn_path)] = wfn_path.read_bytes()
+
+    def propagate_wfn(self, output: SinglePointOutput, prog_inp: ProgramInput) -> None:
+        """Propagate the wavefunction from the previous calculation.
+
+        Args:
+            output: The output from a previous calculation containing wavefunction data.
+            prog_inp: The ProgramInput object on which to place the wavefunction data.
+
+        Returns:
+            None. Modifies the prog_inp object in place.
+        """
+
+        # Naming conventions from TeraChem uses xyz filename as scratch dir postfix
+        scr_postfix = self.xyz_filename.split(".")[0]
+
+        # Wavefunction filenames
+        c0, ca0, cb0 = "c0", "ca0", "cb0"
+
+        c0_bytes = output.files.get(f"scr.{scr_postfix}/{c0}")
+        ca0_bytes = output.files.get(f"scr.{scr_postfix}/{ca0}")
+        cb0_bytes = output.files.get(f"scr.{scr_postfix}/{cb0}")
+
+        if not c0_bytes and not (ca0_bytes and cb0_bytes):
+            raise AdapterInputError(
+                program=self.program,
+                message="Could not find c0 or ca/b0 files in output.",
+            )
+
+        # Load wavefunction data onto ProgramInput object
+
+        if c0_bytes:
+            prog_inp.files[c0] = c0_bytes
+            prog_inp.keywords["guess"] = c0
+
+        else:  # ca0_bytes and cb0_bytes
+            prog_inp.files[ca0] = ca0_bytes
+            prog_inp.files[cb0] = cb0_bytes
+            prog_inp.keywords["guess"] = f"{ca0} {cb0}"