+++ qcop/adapters/utils.py
@@ -2,10 +2,20 @@
 Utility functions for adapters. Avoids circular imports when these functions placed in
 top-level utils.py module.
 """
-
+import logging
+import os
+import platform
+import shutil
 import subprocess
-import time
+import tempfile
+from contextlib import contextmanager
+from io import StringIO
+from pathlib import Path
+from time import time
 from typing import Callable, List, Optional
+
+from qcio import Provenance
+from qcio.helper_types import StrOrPath
 
 from qcop.exceptions import ExternalProgramExecutionError, ProgramNotFoundError
 
@@ -61,7 +71,7 @@ def execute_subprocess(
 
     # Setup variables for monitoring stdout
     stdout_lines: List[str] = []
-    prev_update_time = time.time()
+    prev_update_time = time()
     prev_line_count = len(stdout_lines)
 
     # Read stdout line by line as it is available until proc terminates
@@ -74,14 +84,14 @@ def execute_subprocess(
 
         if (
             update_func
-            and time.time() - prev_update_time > (update_interval or 0.5)
+            and time() - prev_update_time > (update_interval or 0.5)
             and len(stdout_lines) > prev_line_count
         ):
             # Call update_func with the current stdout
             stdout = "".join(stdout_lines)
             stdout_since_last_update = "".join(stdout_lines[prev_line_count:])
             update_func(stdout, stdout_since_last_update)
-            prev_update_time = time.time()
+            prev_update_time = time()
             prev_line_count = len(stdout_lines)
 
     # Unconsumed output may still need to be processed
@@ -97,3 +107,82 @@ def execute_subprocess(
         )
 
     return stdout
+
+
+@contextmanager
+def capture_logs(logger_name: str):
+    """Capture logs from a program during execution."""
+
+    # Create a logger
+    logger = logging.getLogger(logger_name)
+
+    # Set the level to capture all logs
+    logger.setLevel(logging.DEBUG)
+
+    # Create a string buffer and add a StreamHandler using the buffer
+    log_capture_string = StringIO()
+    handler = logging.StreamHandler(log_capture_string)
+
+    # Optional: set a format for the handler
+    formatter = logging.Formatter(
+        "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
+    )
+    handler.setFormatter(formatter)
+
+    # Add the handler to the logger
+    logger.addHandler(handler)
+
+    try:
+        yield (logger, log_capture_string)
+    finally:
+        # Remove the handler when done
+        logger.removeHandler(handler)
+
+
+def construct_provenance(
+    program: str,
+    version: Optional[str],
+    scratch_dir: Path,
+    start: float,
+) -> Provenance:
+    """Construct a provenance object for a calculation.
+
+    Args:
+        program: The program used for the calculation.
+        version: The program version.
+        scratch_dir: The working directory of the calculation.
+        start: The start time of the calculation.
+        stdout: The stdout of the calculation.
+
+    Returns:
+        The Provenance object.
+    """
+    wall_time = time() - start
+
+    return Provenance(
+        program=program,
+        program_version=version,
+        scratch_dir=str(scratch_dir),
+        wall_time=round(wall_time, 6),
+        hostname=platform.node(),
+        hostcpus=os.cpu_count(),
+    )
+
+
+@contextmanager
+def tmpdir(directory: Optional[StrOrPath] = None, rmdir: bool = True):
+    """Context manager for a temporary directory.
+
+    Args:
+        directory: Where to create the temporary directory. If None, a new directory is
+            created in the system default temporary directory.
+        rmdir: Whether to remove the temporary directory when the context manager exits.
+    """
+    temp_dir = Path(directory or tempfile.mkdtemp())  # Set path to directory
+    temp_dir.mkdir(parents=True, exist_ok=True)  # Create directory
+    cwd = Path.cwd()  # Save current working directory
+    os.chdir(temp_dir)  # Change to temporary directory
+    yield temp_dir  # Execute code in context manager
+    if rmdir:  # After exiting context manager
+        shutil.rmtree(temp_dir)
+    os.chdir(cwd)