+++ tests/adaptors/test_file.py
@@ -2,8 +2,8 @@ import pytest
 from qcio import FileInput
 
 from qcop.adapters import FileAdapter
+from qcop.adapters.utils import tmpdir
 from qcop.exceptions import ProgramNotFoundError
-from qcop.utils import tmpdir
 
 
 def test_file_adapter_compute(tmp_path):
@@ -12,7 +12,7 @@ def test_file_adapter_compute(tmp_path):
     file_inp.files["hello_world.py"] = "print('hello world')"
     with tmpdir() as path:
         file_inp.write_files(path)
-        _, stdout = FileAdapter("python").compute(file_inp)
+        _, stdout = FileAdapter("python").compute_results(file_inp)
     assert stdout == "hello world\n"
 
 
@@ -20,4 +20,4 @@ def test_file_adapter_raises_exception_if_program_fails():
     # Create FileInput
     file_inp = FileInput()
     with pytest.raises(ProgramNotFoundError):
-        FileAdapter("does_not_exist").compute(file_inp)
+        FileAdapter("does_not_exist").compute_results(file_inp)